Combine multiple ReadGenesis functions.
